+++ src/vulnix/main.py
@@ -21,37 +21,19 @@ See vulnix --help for a full list of options.
 from .nix import Store
 from .nvd import NVD, DEFAULT_MIRROR, DEFAULT_CACHE_DIR
 from .whitelist import WhiteList
+from .utils import cve_url, Timer
 import click
-import glob
+import json
 import logging
-import os
-import os.path
 import pkg_resources
 import sys
-import time
 import urllib.request
 
 DEFAULT_WHITELIST = pkg_resources.resource_filename(
     __name__, 'default_whitelist.yaml')
+CURRENT_SYSTEM = '/nix/var/nix/gcroots/current-system'
 
 _log = logging.getLogger(__name__)
-
-
-class Timer:
-
-    def __init__(self, debugmsg):
-        self.debugmsg = debugmsg
-
-    def __enter__(self):
-        _log.debug('Starting: ' + self.debugmsg)
-        self.start = time.clock()
-        return self
-
-    def __exit__(self, *args):
-        self.end = time.clock()
-        self.interval = self.end - self.start
-        _log.debug('Finished: {} took {:.2f} seconds'.format(
-                   self.debugmsg, self.interval))
 
 
 def howto():
@@ -60,7 +42,36 @@ def howto():
     click.echo(tail, nl=False)
 
 
-def output(affected_derivations, verbosity):
+def init_logging(verbose, debug):
+    if debug:
+        logging.basicConfig(level=logging.DEBUG)
+    else:
+        logging.getLogger('requests').setLevel(logging.ERROR)
+        if verbose:
+            logging.basicConfig(level=logging.INFO)
+        else:
+            logging.basicConfig(level=logging.WARNING)
+
+
+def output_json(affected_derivations):
+    out = []
+    status = [0]
+    # get rid of duplicate entries
+    for d in {deriv.name: deriv for deriv in affected_derivations}.values():
+        out.append({
+            'name': d.name,
+            'pname': d.pname,
+            'version': d.version,
+            'whitelist': d.status,
+            'derivation': d.store_path,
+            'affected_by': list(d.affected_by),
+        })
+        status.append(1 if d.status == 'inprogress' else 2)
+    print(json.dumps(out, indent=1))
+    return max(status)
+
+
+def output(affected_derivations, verbosity, notfixed):
     status = []
     derivations = []
     seen = {}
@@ -71,17 +82,28 @@ def output(affected_derivations, verbosity):
             continue
         seen[marker] = 1
         derivations.append(item)
-    derivations.sort(key=lambda k: k.simple_name)
+    derivations.sort(key=lambda k: k.pname)
 
     amount = len(derivations)
-    names = ', '.join(derivations[k].simple_name for k in range(3))
-    summary = 'Found {} advisories for {}'.format(amount, names)
-    if amount > 3:
-        summary += ', ... (and {:d} more)'.format(amount - 3)
+    if amount == 0:
+        summary = 'Found no advisories'
+    else:
+        names = ', '.join(d.pname for d in derivations[:3])
+        summary = 'Found {} advisories for {}'.format(amount, names)
+        if amount > 3:
+            summary += ', ... (and {:d} more)'.format(amount - 3)
     click.secho(summary, fg='red')
 
     for derivation in derivations:
-        progress = '*' if derivation.status == 'inprogress' else ''
+        if derivation.status == 'inprogress':
+            progress = '*'
+        elif derivation.status == 'notfixed':
+            # Don't show 'notfixed' derivations if not wanted.
+            if not notfixed:
+                continue
+            progress = '!'
+        else:
+            progress = ''
         click.echo('\n{}\n{}{}\n'.format('=' * 72, derivation.name, progress))
         if verbosity >= 1:
             click.echo(derivation.store_path)
@@ -94,27 +116,24 @@ def output(affected_derivations, verbosity):
                 for root in derivation.roots():
                     click.echo("\t" + root)
         click.echo("CVEs:")
-        for cve in derivation.affected_by:
-            click.echo("\t" + cve.url)
+        for cve_id in derivation.affected_by:
+            click.echo("\t" + cve_url(cve_id))
         status.append(1 if derivation.status == 'inprogress' else 2)
 
     return max(status)
 
 
-def populate_store(gc_roots, system, paths):
+def populate_store(gc_roots, paths, requisites=True):
     """Load derivations from nix store depending on cmdline invocation."""
-    store = Store()
+    store = Store(requisites)
     if gc_roots:
         store.add_gc_roots()
-    if system:
-        store.add_path('/nix/var/nix/gcroots/current-system')
-    if paths:
-        for path in paths:
-            store.add_path(path)
+    for path in paths:
+        store.add_path(path)
     return store
 
 
-class Ressource:
+class Resource:
 
     def __init__(self, url):
         self.url = url
@@ -129,106 +148,109 @@ class Ressource:
             self.fp = open(url)
 
 
-def open_ressource(ctx, param, value):
+def open_resource(ctx, param, value):
     """returns fp for files or remote ressources"""
     if value:
         for v in value:
-            yield Ressource(v)
+            yield Resource(v)
+
+
+def load_whitelists(default_whitelist, whitelist):
+    wl = WhiteList()
+    if default_whitelist:
+        with open(DEFAULT_WHITELIST) as f:
+            wl.parse(f)
+    if whitelist:
+        for arg in whitelist:
+            wl.parse(arg.fp)
+    return wl
+
+
+def run(nvd, store, whitelist):
+    affected = set()
+    for derivation in store.derivations.values():
+        derivation.check(nvd)
+        if derivation.is_affected:
+            affected.add(derivation)
+    return affected
 
 
 @click.command('vulnix')
+# what to scan
 @click.option('-S', '--system', is_flag=True,
               help='Scan the current system')
 @click.option('-G', '--gc-roots', is_flag=True,
               help='Scan all active GC roots (including old ones)')
-@click.option('-w', '--whitelist', multiple=True, callback=open_ressource,
+@click.argument('path', nargs=-1, type=click.Path(exists=True))
+# modify operation
+@click.option('-w', '--whitelist', multiple=True, callback=open_resource,
               help='Add another whitelist ressource to declare exceptions.')
-@click.option('-m', '--mirror',
-              help='Mirror to fetch NVD archives from. Default: {}'.format(
-                  DEFAULT_MIRROR),
-              default=DEFAULT_MIRROR)
 @click.option('--default-whitelist/--no-default-whitelist', default=True,
               help='Load built-in base whitelist from "{}". Additional '
               'whitelist files can be specified using the "-w" option. '
               'Default: yes'.format(DEFAULT_WHITELIST))
+@click.option('-c', '--cache-dir', type=click.Path(file_okay=False),
+              default=DEFAULT_CACHE_DIR,
+              help='Cache directory to store parsed archive data. '
+              'Default: {}'.format(DEFAULT_CACHE_DIR))
+@click.option('-r/-R', '--requisites/--no-requisites', default=True,
+              help='Determine transitive closure vs. just examine passed '
+              'derivations (default: yes)')
+@click.option('-m', '--mirror',
+              help='Mirror to fetch NVD archives from. Default: {}'.format(
+                  DEFAULT_MIRROR),
+              default=DEFAULT_MIRROR)
+# output control
 @click.option('-d', '--debug', is_flag=True,
               help='Show debug information.')
 @click.option('-v', '--verbose', count=True,
               help='Increase output verbosity.')
-@click.option('-c', '--cache-dir', default=DEFAULT_CACHE_DIR,
-              help='Cache directory to store parsed archive data. '
-              'Default: {}'.format(DEFAULT_CACHE_DIR))
-@click.option('-U', '--update-cache', is_flag=True,
-              help='Update the parsed archives and exit')
 @click.option('-V', '--version', is_flag=True,
               help='Print vulnix version and exit')
-@click.argument('path', nargs=-1,
-                type=click.Path(exists=True))
-def main(debug, verbose, whitelist, default_whitelist,
-         gc_roots, system, path, mirror, cache_dir, update_cache, version):
-    """Scans nix store paths for derivations with security vulnerabilities."""
-    if debug:
-        logging.basicConfig(level=logging.DEBUG)
-    else:
-        logging.getLogger('requests').setLevel(logging.ERROR)
-        if verbose:
-            logging.basicConfig(level=logging.INFO)
-        else:
-            logging.basicConfig(level=logging.WARNING)
-
+@click.option('-F', '--notfixed', is_flag=True,
+              help='Show packages which are not fixed by upstream')
+@click.option('-j', '--json/--no-json', help='JSON vs. human readable output')
+def main(debug, verbose, whitelist, default_whitelist, gc_roots, system, path,
+         mirror, cache_dir, version, notfixed, requisites, json):
     if version:
         print('vulnix ' + pkg_resources.get_distribution('vulnix').version)
         sys.exit(0)
 
-    if not (update_cache or gc_roots or system or path):
+    if not (gc_roots or system or path):
         howto()
         sys.exit(3)
 
-    cache_dir = os.path.expanduser(cache_dir)
+    init_logging(verbose, debug)
 
-    if not os.path.exists(cache_dir):
-        os.makedirs(cache_dir)
+    paths = list(path)
+    if system:
+        paths.append(CURRENT_SYSTEM)
 
-    # Clean up old-style data
-    for file in glob.glob(cache_dir+'/*.xml*'):
-        os.unlink(file)
-
-    nvd = NVD(mirror=mirror, cache_dir=cache_dir)
-    with nvd:
-        with Timer('Load NVD data'):
-            nvd.update()
-            if update_cache:
-                sys.exit(0)
-
+    try:
         with Timer('Load whitelist'):
-            wl = WhiteList()
-            if default_whitelist:
-                with open(DEFAULT_WHITELIST) as f:
-                    wl.parse(f)
-            if whitelist:
-                for res in whitelist:
-                    wl.parse(res.fp)
+            wl = load_whitelists(default_whitelist, whitelist)
 
         with Timer('Load derivations'):
-            store = populate_store(gc_roots, system, path)
+            store = populate_store(gc_roots, paths, requisites)
 
-        affected = set()
-        with Timer('Scan vulnerabilities'):
-            for derivation in store.derivations.values():
-                with Timer('Scan {}'.format(derivation.name)):
-                    derivation.check(nvd, wl)
-                    if derivation.is_affected:
-                        affected.add(derivation)
+        nvd = NVD(mirror, cache_dir)
+        with nvd:
+            with Timer('Load NVD data'):
+                nvd.update()
+            with Timer('Scan vulnerabilities'):
+                affected = run(nvd, store, wl)
 
-        returncode = 0
+        if json:
+            sys.exit(output_json(affected))
         if affected:
             # sensu maps following return codes
             # 0 - ok, 1 - warning, 2 - critical, 3 - unknown
-            returncode = output(affected, verbose)
+            sys.exit(output(affected, verbose, notfixed))
         else:
             click.secho('vulnix: no vulnerabilities detected', fg='green')
-            _log.debug('returncode %d', returncode)
 
     # This needs to happen outside the NVD context: otherwise ZODB will abort
     # the transaction and we will keep updating over and over.
-    sys.exit(returncode)
+    except RuntimeError as e:
+        _log.critical(e)
+        sys.exit(2)