+++ src/vulnix/tests/test_nvd.py
@@ -1,4 +1,5 @@
 from vulnix.nvd import NVD
+from vulnix.utils import cve_url
 import http.server
 import os
 import pytest
@@ -15,7 +16,6 @@ def http_server():
     t = threading.Thread(target=httpd.serve_forever, daemon=True)
     t.start()
     yield mirror_url
-    httpd.shutdown()
 
 
 def test_update_and_parse(tmpdir, http_server):
@@ -24,14 +24,14 @@ def test_update_and_parse(tmpdir, http_server):
     with nvd:
         nvd.update()
         modified = nvd._root['archives']['Modified']
-        assert len(modified.products) == 6
+        assert len(modified.products) == 7
 
         mariadb = modified.products['mariadb']
         mariadb = list(sorted(mariadb, key=lambda x: x.cve_id))
         cve = mariadb[0]
         assert cve.cve_id == 'CVE-2016-6664'
         assert (
-            cve.url ==
+            cve_url(cve.cve_id) ==
             'https://web.nvd.nist.gov/view/vuln/detail?vulnId=CVE-2016-6664')
         cpe = list(sorted(cve.affected_products, key=lambda x: x.vendor))[1]
         assert cpe.versions == {'5.7.14', '5.5.51', '5.6.32'}