+++ src/vulnix/tests/test_whitelist.py
@@ -1,21 +1,18 @@
-from vulnix.whitelist import WhiteList
-import os.path as p
+from pkg_resources import resource_stream
 import pytest
+
+from vulnix.whitelist import WhiteList, WhiteListRule
 
 
 @pytest.yield_fixture
 def test_whitelist():
-    fn = p.join(p.dirname(__file__), 'test_whitelist.yaml')
-    with open(fn) as f:
-        yield f
+    return resource_stream('vulnix', 'tests/fixtures/test_whitelist.yaml')
 
 
 def test_scan_rulefile(test_whitelist):
     w = WhiteList()
     w.parse(test_whitelist)
-    assert len(w.rules) == 7  # list of CVEs count for each cve_id
-
-    r = w.rules
+    assert len(w.rules) == 8  # list of CVEs count for each cve_id
 
     r = w.rules.pop(0)
     assert r.cve == 'CVE-2015-2504'
@@ -24,6 +21,8 @@ def test_scan_rulefile(test_whitelist):
     assert r.comment is None
     assert r.vendor is None
     assert r.product is None
+    assert r.status == 'ignore'
+    assert str(r) == r.cve
 
     r = w.rules.pop(0)
     assert r.cve == 'CVE-2015-7696'
@@ -32,18 +31,21 @@ def test_scan_rulefile(test_whitelist):
     assert r.comment is None
     assert r.vendor is None
     assert r.product is None
+    assert r.status is 'ignore'
+    assert str(r) == r.cve
 
     r = w.rules.pop(0)
     assert r.cve == 'CVE-2015-2503'
     assert r.name is None
     assert r.version is None
-    print(r.comment)
     assert r.comment == """microsoft access, accidentally matching the 'access' derivation
 
 https://plan.flyingcircus.io/issues/18544
 """
     assert r.vendor is None
     assert r.product is None
+    assert r.status is 'ignore'
+    assert str(r) == r.cve
 
     r = w.rules.pop(0)
     assert r.cve is None
@@ -52,6 +54,8 @@ https://plan.flyingcircus.io/issues/18544
     assert r.comment is None
     assert r.vendor is None
     assert r.product is None
+    assert r.status is 'ignore'
+    assert str(r) == r.name
 
     r = w.rules.pop(0)
     assert r.cve == 'CVE-2015-7696'
@@ -60,6 +64,8 @@ https://plan.flyingcircus.io/issues/18544
     assert r.comment is None
     assert r.vendor is None
     assert r.product is None
+    assert r.status == 'inprogress'
+    assert str(r) == '{} ({})'.format(r.name, r.cve)
 
     r = w.rules.pop(0)
     assert r.cve is None
@@ -68,6 +74,8 @@ https://plan.flyingcircus.io/issues/18544
     assert r.comment is None
     assert r.vendor is None
     assert r.product is None
+    assert r.status is 'ignore'
+    assert str(r) == '{}-{}'.format(r.name, r.version)
 
     r = w.rules.pop(0)
     assert r.cve is None
@@ -76,13 +84,30 @@ https://plan.flyingcircus.io/issues/18544
     assert r.comment is None
     assert r.vendor == 'microsoft'
     assert r.product == 'access'
+    assert r.status is 'ignore'
+    assert str(r) == '{}-{}'.format(r.vendor, r.product)
 
 
 def test_concatenate_multiple_whitelists(test_whitelist):
     w = WhiteList()
     w.parse(test_whitelist)
-    with open(p.join(p.dirname(__file__), 'test_whitelist2.yaml')) as f:
+    with resource_stream('vulnix', 'tests/fixtures/test_whitelist2.yaml') as f:
         w.parse(f)
 
-    assert len(w.rules) == 8  # combined list
+    assert len(w.rules) == 9  # combined list
     assert w.rules[-1].cve == 'CVE-2016-0001'
+
+
+def test_no_matchable_attribute():
+    with pytest.raises(RuntimeError):
+        WhiteListRule()
+
+
+def test_invalid_status():
+    with pytest.raises(RuntimeError):
+        WhiteListRule(name='libxslt', version='1.2.3', status='foobar')
+
+
+def test_version_must_be_string():
+    with pytest.raises(RuntimeError):
+        WhiteListRule(version=1)