+++ CHANGES.rst
@@ -1,10 +1,60 @@
 Changes
 =======
 
-1.2.3 (unreleased)
+
+1.4.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Ignore case when guessing CVE identifiers from patch file names (thanks to
+  adisbladis).
+
+
+1.4.0 (2017-11-27)
+------------------
+
+- Guesses applied CVE patches out of the `patches` derivation envVar (see
+  nixpkgs #15660).
+
+
+1.3.4 (2017-10-29)
+------------------
+
+- Add '--no-requisites' flag which stops vulnix from determining the transitive
+  closure of derivations passed on the command line.
+- Provide structured JSON output with `--json`.
+- Remove whitelist from README as it is quite buggy right now.
+
+
+1.3.3 (2017-10-16)
+------------------
+
+- Fix return code bug (#28741).
+- Fix partial whitelisting of products where several vulnerable versions are
+  present on the system at the same time (#24).
+- Improve error reporting for incorrectly formed whitelist rules.
+
+
+1.3.2 (2017-10-06)
+------------------
+
+- Minor: fix packaging issues.
+
+
+1.3.1 (2017-10-06)
+------------------
+
+- Security: Fix arbitrary code execution bug during derivation evaluation.
+
+
+1.3.0 (2017-09-18)
+------------------
+
+- `.drv` files may be specified directly on the command line.
+- Updated PyPI dependencies.
+- Document system requirements (#12).
+- Don't leave large files in /tmp around.
+- Remove duplicate CVEs in output (#25).
+- Fix bug with reporting less than 3 vulnerabilities (#28).
 
 
 1.2.2 (2017-01-28)